download_worker/requirements.txt
redis
pytube
mysql.connector
uuid
dotenv